requirements.txt
spotipy
pandas
scikit-learn
gradio
Pillow
stability-sdk
protobuf
streamlit